Removing global variable for app.
The routes aren't being replaced, so let's just add the engine in all the time. Avoids the need to rewrite routes and hold onto the app.
