+++ spec/fake_app/rails_app.rb
@@ -5,22 +5,33 @@ require 'action_view/railtie'
 require 'fake_app/active_record/config' if defined? ActiveRecord
 require 'fake_app/mongoid/config' if defined? Mongoid
 # config
-$app = Class.new(Rails::Application)
-$app.config.secret_token = '3b7cd727ee24e8444053437c36cc66c4'
-$app.config.session_store :cookie_store, :key => '_myapp_session'
-$app.config.active_support.deprecation = :log
-$app.config.eager_load = false
+app = Class.new(Rails::Application)
+app.config.secret_token = '3b7cd727ee24e8444053437c36cc66c4'
+app.config.session_store :cookie_store, :key => '_myapp_session'
+app.config.active_support.deprecation = :log
+app.config.eager_load = false
 # Rais.root
-$app.config.root = File.dirname(__FILE__)
+app.config.root = File.dirname(__FILE__)
 Rails.backtrace_cleaner.remove_silencers!
-$app.initialize!
+app.initialize!
+
+module MyEngine
+  class Engine < ::Rails::Engine
+    isolate_namespace MyEngine
+    routes.draw do
+      resources :items
+    end
+  end
+end
 
 # routes
-$app.routes.draw do
+app.routes.draw do
   resources :users
   resources :addresses do
     get 'page/:page', :action => :index, :on => :collection
   end
+
+  mount MyEngine::Engine, :at => '/mounted_engine'
 end
 
 #models
